even less of a wip

+++ src/url.c
@@ -13,9 +13,6 @@
 
 #include "natsp.h"
 #include "util.h"
-
-#include <string.h>
-
 #include "mem.h"
 
 void
@@ -37,7 +34,7 @@ _parsePort(int *port, const char *sport)
     natsStatus  s    = NATS_OK;
     int64_t     n    = 0;
 
-    n = nats_ParseInt64(sport, (int) strlen(sport));
+    n = nats_ParseInt64(sport, strlen(sport));
     if ((n < 0) || (n > INT32_MAX))
         s = nats_setError(NATS_INVALID_ARG, "invalid port '%s'", sport);
     else
@@ -49,21 +46,21 @@ _parsePort(int *port, const char *sport)
 natsStatus
 natsUrl_Create(natsUrl **newUrl, const char *urlStr)
 {
-    natsStatus  s      = NATS_OK;
-    char        *copy  = NULL;
-    char        *ptr   = NULL;
-    const char  *scheme= NULL;
-    const char  *user  = NULL;
-    const char  *pwd   = NULL;
-    const char  *host  = NULL;
-    const char  *port  = NULL;
-    const char  *path  = NULL;
-    natsUrl     *url   = NULL;
+    natsStatus s = NATS_OK;
+    char *copy = NULL;
+    char *ptr = NULL;
+    const char *scheme = NULL;
+    const char *user = NULL;
+    const char *pwd = NULL;
+    const char *host = NULL;
+    const char *port = NULL;
+    const char *path = NULL;
+    natsUrl *url = NULL;
 
     if (nats_IsStringEmpty(urlStr))
         return nats_setDefaultError(NATS_INVALID_ARG);
 
-    url = (natsUrl*) NATS_CALLOC(1, sizeof(natsUrl));
+    url = (natsUrl *)NATS_CALLOC(1, sizeof(natsUrl));
     if (url == NULL)
         return nats_setDefaultError(NATS_NO_MEMORY);
 
@@ -81,18 +78,18 @@ natsUrl_Create(natsUrl **newUrl, const char *urlStr)
         else
         {
             *ptr = '\0';
-            scheme = (const char*) copy;
+            scheme = (const char *)copy;
             ptr += 3;
         }
     }
     // User info
     if (s == NATS_OK)
     {
-        char *sep  = strrchr(ptr, '@');
+        char *sep = strrchr(ptr, '@');
 
         if (sep != NULL)
         {
-            host = (const char*) (sep+1);
+            host = (const char *)(sep + 1);
             *sep = '\0';
 
             if (ptr != sep)
@@ -102,19 +99,19 @@ natsUrl_Create(natsUrl **newUrl, const char *urlStr)
                 {
                     *sep = '\0';
                     if (sep != ptr)
-                        user = (const char*) ptr;
-                    if (sep+1 != host)
-                        pwd = (const char*) (sep+1);
+                        user = (const char *)ptr;
+                    if (sep + 1 != host)
+                        pwd = (const char *)(sep + 1);
                 }
                 else
                 {
-                    user = (const char*) ptr;
+                    user = (const char *)ptr;
                 }
             }
         }
         else
         {
-            host = (const char*) ptr;
+            host = (const char *)ptr;
         }
     }
     // Host
@@ -123,14 +120,14 @@ natsUrl_Create(natsUrl **newUrl, const char *urlStr)
         // Search for end of IPv6 address (if applicable)
         ptr = strrchr(host, ']');
         if (ptr == NULL)
-            ptr = (char*) host;
+            ptr = (char *)host;
 
         // From that point, search for the last ':' character
         ptr = strrchr(ptr, ':');
         if (ptr != NULL)
         {
             *ptr = '\0';
-            port = (const char*) (ptr+1);
+            port = (const char *)(ptr + 1);
         }
         if (nats_IsStringEmpty(host))
             host = "localhost";
@@ -145,7 +142,7 @@ natsUrl_Create(natsUrl **newUrl, const char *urlStr)
             if (sep != NULL)
             {
                 *sep = '\0';
-                path = (const char*) (sep+1);
+                path = (const char *)(sep + 1);
             }
         }
         if (nats_IsStringEmpty(port))
@@ -156,12 +153,12 @@ natsUrl_Create(natsUrl **newUrl, const char *urlStr)
     // Assemble everything
     if (s == NATS_OK)
     {
-        const char  *userval    = (nats_IsStringEmpty(user) ? "" : user);
-        const char  *usep       = (nats_IsStringEmpty(pwd) ? "" : ":");
-        const char  *pwdval     = (nats_IsStringEmpty(pwd) ? "" : pwd);
-        const char  *hsep       = (nats_IsStringEmpty(user) ? "" : "@");
-        const char  *pathsep    = (nats_IsStringEmpty(path) ? "" : "/");
-        const char  *pathval    = (nats_IsStringEmpty(path) ? "" : path);
+        const char *userval = (nats_IsStringEmpty(user) ? "" : user);
+        const char *usep = (nats_IsStringEmpty(pwd) ? "" : ":");
+        const char *pwdval = (nats_IsStringEmpty(pwd) ? "" : pwd);
+        const char *hsep = (nats_IsStringEmpty(user) ? "" : "@");
+        const char *pathsep = (nats_IsStringEmpty(path) ? "" : "/");
+        const char *pathval = (nats_IsStringEmpty(path) ? "" : path);
 
         DUP_STRING(s, url->host, host);
 
@@ -171,7 +168,7 @@ natsUrl_Create(natsUrl **newUrl, const char *urlStr)
             IF_OK_DUP_STRING(s, url->password, pwd);
 
         if ((s == NATS_OK) && nats_asprintf(&url->fullUrl, "%s://%s%s%s%s%s:%d%s%s",
-                scheme, userval, usep, pwdval, hsep, host, url->port, pathsep, pathval) < 0)
+                                            scheme, userval, usep, pwdval, hsep, host, url->port, pathsep, pathval) < 0)
         {
             s = nats_setDefaultError(NATS_NO_MEMORY);
         }
@@ -186,3 +183,144 @@ natsUrl_Create(natsUrl **newUrl, const char *urlStr)
 
     return NATS_UPDATE_ERR_STACK(s);
 }
+
+// natsStatus
+// natsUrl_Create(natsUrl **newUrl, const char *urlStr)
+// {
+//     natsStatus s = NATS_OK;
+//     uint8_t *copy = NULL;
+//     uint8_t *ptr = NULL;
+//     uint8_t *scheme = NULL;
+//     uint8_t *user = NULL;
+//     uint8_t *pwd = NULL;
+//     uint8_t *host = NULL;
+//     uint8_t *port = NULL;
+//     uint8_t *path = NULL;
+//     natsUrl *url = NULL;
+
+//     if (nats_IsStringEmpty(urlStr))
+//         return nats_setDefaultError(NATS_INVALID_ARG);
+
+//     url = (natsUrl*) NATS_CALLOC(1, sizeof(natsUrl));
+//     if (url == NULL)
+//         return nats_setDefaultError(NATS_NO_MEMORY);
+
+//     s = nats_Trim(&copy, (const uint8_t*)urlStr);
+
+//     // Scheme
+//     if (s == NATS_OK)
+//     {
+//         ptr = nats_strstr(copy, "://");
+//         if (ptr == NULL)
+//         {
+//             scheme = ()"nats";
+//             ptr = copy;
+//         }
+//         else
+//         {
+//             *ptr = '\0';
+//             scheme = copy;
+//             ptr += 3;
+//         }
+//     }
+//     // User info
+//     if (s == NATS_OK)
+//     {
+//         uint8_t *sep  = nats_strrchr(ptr, '@');
+
+//         if (sep != NULL)
+//         {
+//             host = sep+1;
+//             *sep = '\0';
+
+//             if (ptr != sep)
+//             {
+//                 sep = nats_strchr(ptr, ':');
+//                 if (sep != NULL)
+//                 {
+//                     *sep = '\0';
+//                     if (sep != ptr)
+//                         user = (const uint8_t*) ptr;
+//                     if (sep+1 != host)
+//                         pwd = (const uint8_t *)(sep + 1);
+//                 }
+//                 else
+//                 {
+//                     user = (const uint8_t *)ptr;
+//                 }
+//             }
+//         }
+//         else
+//         {
+//             host = (const uint8_t*) ptr;
+//         }
+//     }
+//     // Host
+//     if (s == NATS_OK)
+//     {
+//         // Search for end of IPv6 address (if applicable)
+//         ptr = nats_strrchr(host, ']');
+//         if (ptr == NULL)
+//             ptr = (char*) host;
+
+//         // From that point, search for the last ':' character
+//         ptr = nats_strrchr(ptr, ':');
+//         if (ptr != NULL)
+//         {
+//             *ptr = '\0';
+//             port = (const uint8_t*) (ptr+1);
+//         }
+//         if (nats_IsStringEmpty(host))
+//             host = "localhost";
+//     }
+//     // Port
+//     if (s == NATS_OK)
+//     {
+//         if (port != NULL)
+//         {
+//             char *sep = strchr(port, '/');
+
+//             if (sep != NULL)
+//             {
+//                 *sep = '\0';
+//                 path = (const char*) (sep+1);
+//             }
+//         }
+//         if (nats_IsStringEmpty(port))
+//             url->port = 4222;
+//         else
+//             s = _parsePort(&url->port, port);
+//     }
+//     // Assemble everything
+//     if (s == NATS_OK)
+//     {
+//         const char  *userval    = (nats_IsStringEmpty(user) ? "" : user);
+//         const char  *usep       = (nats_IsStringEmpty(pwd) ? "" : ":");
+//         const char  *pwdval     = (nats_IsStringEmpty(pwd) ? "" : pwd);
+//         const char  *hsep       = (nats_IsStringEmpty(user) ? "" : "@");
+//         const char  *pathsep    = (nats_IsStringEmpty(path) ? "" : "/");
+//         const char  *pathval    = (nats_IsStringEmpty(path) ? "" : path);
+
+//         DUP_STRING(s, url->host, host);
+
+//         if (user != NULL)
+//             IF_OK_DUP_STRING(s, url->username, user);
+//         if (pwd != NULL)
+//             IF_OK_DUP_STRING(s, url->password, pwd);
+
+//         if ((s == NATS_OK) && nats_asprintf(&url->fullUrl, "%s://%s%s%s%s%s:%d%s%s",
+//                 scheme, userval, usep, pwdval, hsep, host, url->port, pathsep, pathval) < 0)
+//         {
+//             s = nats_setDefaultError(NATS_NO_MEMORY);
+//         }
+//     }
+
+//     NATS_FREE(copy);
+
+//     if (s == NATS_OK)
+//         *newUrl = url;
+//     else
+//         natsUrl_Destroy(url);
+
+//     return NATS_UPDATE_ERR_STACK(s);
+// }
